suggestions for typo fixes
Do still checkout this chapter for formatting: \- <c>tests/test_default.py</c>: Basic Testinfra test, can be extended. If the you are not going to write any Python tests, the "molecule/<scenario>/tests" directory should be removed entirely and have the test disabled in **molecule.yml**.

It seems to drop scenario, and molecule is not bolded.